+++ ta-lib-main/ta_codegen/output/java/Core_STOCH.java
@@ -253,6 +253,9 @@
        *  caller buffer because more input data then the
        *  requested range was needed for doing %D).
        */
+      /* memmove, not memcpy: tempBuffer aliases outSlowK when the caller buffer is
+       * reused as scratch, so source and destination overlap (issue #94).
+       */
       System.arraycopy(tempBuffer, lookbackDSlow, outSlowK, 0, (int)outNBElement.value * 1);
       /* Don't need K anymore, free it if it was allocated here. */
       if( (bufferIsAllocated) != 0 ) {
@@ -478,13 +481,13 @@
          tempBuffer = new double[(int)((endIdx - today + 1) * 1)];
       }
       while( today <= endIdx ) {
-         tmp = inLow[today];
+         tmp = (double)inLow[today];
          if( lowestIdx < trailingIdx ) {
             lowestIdx = trailingIdx;
-            lowest = inLow[lowestIdx];
+            lowest = (double)inLow[lowestIdx];
             i = lowestIdx;
             while( ++i <= today ) {
-               tmp = inLow[i];
+               tmp = (double)inLow[i];
                if( tmp < lowest ) {
                   lowestIdx = i;
                   lowest = tmp;
@@ -496,13 +499,13 @@
             lowest = tmp;
             diff = (highest - lowest) / 100.0;
          }
-         tmp = inHigh[today];
+         tmp = (double)inHigh[today];
          if( highestIdx < trailingIdx ) {
             highestIdx = trailingIdx;
-            highest = inHigh[highestIdx];
+            highest = (double)inHigh[highestIdx];
             i = highestIdx;
             while( ++i <= today ) {
-               tmp = inHigh[i];
+               tmp = (double)inHigh[i];
                if( tmp > highest ) {
                   highestIdx = i;
                   highest = tmp;
@@ -515,7 +518,7 @@
             diff = (highest - lowest) / 100.0;
          }
          if( diff != 0.0 ) {
-            tempBuffer[outIdx++] = (inClose[today] - lowest) / diff;
+            tempBuffer[outIdx++] = ((double)inClose[today] - lowest) / diff;
          } else {
             tempBuffer[outIdx++] = 0.0;
          }
@@ -604,13 +607,13 @@
          tempBuffer = new double[(int)((endIdx - today + 1) * 1)];
       }
       while( today <= endIdx ) {
-         tmp = inLow[today];
+         tmp = (double)inLow[today];
          if( lowestIdx < trailingIdx ) {
             lowestIdx = trailingIdx;
-            lowest = inLow[lowestIdx];
+            lowest = (double)inLow[lowestIdx];
             i = lowestIdx;
             while( ++i <= today ) {
-               tmp = inLow[i];
+               tmp = (double)inLow[i];
                if( tmp < lowest ) {
                   lowestIdx = i;
                   lowest = tmp;
@@ -622,13 +625,13 @@
             lowest = tmp;
             diff = (highest - lowest) / 100.0;
          }
-         tmp = inHigh[today];
+         tmp = (double)inHigh[today];
          if( highestIdx < trailingIdx ) {
             highestIdx = trailingIdx;
-            highest = inHigh[highestIdx];
+            highest = (double)inHigh[highestIdx];
             i = highestIdx;
             while( ++i <= today ) {
-               tmp = inHigh[i];
+               tmp = (double)inHigh[i];
                if( tmp > highest ) {
                   highestIdx = i;
                   highest = tmp;
@@ -641,7 +644,7 @@
             diff = (highest - lowest) / 100.0;
          }
          if( diff != 0.0 ) {
-            tempBuffer[outIdx++] = (inClose[today] - lowest) / diff;
+            tempBuffer[outIdx++] = ((double)inClose[today] - lowest) / diff;
          } else {
             tempBuffer[outIdx++] = 0.0;
          }
