+++ ta-lib-main/ta_codegen/output/java/Core_STOCHF.java
@@ -232,6 +232,9 @@
        *  caller buffer because more input data then the
        *  requested range was needed for doing %D).
        */
+      /* memmove, not memcpy: tempBuffer aliases outFastK when the caller buffer is
+       * reused as scratch, so source and destination overlap (issue #94).
+       */
       System.arraycopy(tempBuffer, lookbackFastD, outFastK, 0, (int)outNBElement.value * 1);
       /* Don't need K anymore, free it if it was allocated here. */
       if( (bufferIsAllocated) != 0 ) {
@@ -443,13 +446,13 @@
          tempBuffer = new double[(int)((endIdx - today + 1) * 1)];
       }
       while( today <= endIdx ) {
-         tmp = inLow[today];
+         tmp = (double)inLow[today];
          if( lowestIdx < trailingIdx ) {
             lowestIdx = trailingIdx;
-            lowest = inLow[lowestIdx];
+            lowest = (double)inLow[lowestIdx];
             i = lowestIdx;
             while( ++i <= today ) {
-               tmp = inLow[i];
+               tmp = (double)inLow[i];
                if( tmp < lowest ) {
                   lowestIdx = i;
                   lowest = tmp;
@@ -461,13 +464,13 @@
             lowest = tmp;
             diff = (highest - lowest) / 100.0;
          }
-         tmp = inHigh[today];
+         tmp = (double)inHigh[today];
          if( highestIdx < trailingIdx ) {
             highestIdx = trailingIdx;
-            highest = inHigh[highestIdx];
+            highest = (double)inHigh[highestIdx];
             i = highestIdx;
             while( ++i <= today ) {
-               tmp = inHigh[i];
+               tmp = (double)inHigh[i];
                if( tmp > highest ) {
                   highestIdx = i;
                   highest = tmp;
@@ -480,7 +483,7 @@
             diff = (highest - lowest) / 100.0;
          }
          if( diff != 0.0 ) {
-            tempBuffer[outIdx++] = (inClose[today] - lowest) / diff;
+            tempBuffer[outIdx++] = ((double)inClose[today] - lowest) / diff;
          } else {
             tempBuffer[outIdx++] = 0.0;
          }
@@ -564,13 +567,13 @@
          tempBuffer = new double[(int)((endIdx - today + 1) * 1)];
       }
       while( today <= endIdx ) {
-         tmp = inLow[today];
+         tmp = (double)inLow[today];
          if( lowestIdx < trailingIdx ) {
             lowestIdx = trailingIdx;
-            lowest = inLow[lowestIdx];
+            lowest = (double)inLow[lowestIdx];
             i = lowestIdx;
             while( ++i <= today ) {
-               tmp = inLow[i];
+               tmp = (double)inLow[i];
                if( tmp < lowest ) {
                   lowestIdx = i;
                   lowest = tmp;
@@ -582,13 +585,13 @@
             lowest = tmp;
             diff = (highest - lowest) / 100.0;
          }
-         tmp = inHigh[today];
+         tmp = (double)inHigh[today];
          if( highestIdx < trailingIdx ) {
             highestIdx = trailingIdx;
-            highest = inHigh[highestIdx];
+            highest = (double)inHigh[highestIdx];
             i = highestIdx;
             while( ++i <= today ) {
-               tmp = inHigh[i];
+               tmp = (double)inHigh[i];
                if( tmp > highest ) {
                   highestIdx = i;
                   highest = tmp;
@@ -601,7 +604,7 @@
             diff = (highest - lowest) / 100.0;
          }
          if( diff != 0.0 ) {
-            tempBuffer[outIdx++] = (inClose[today] - lowest) / diff;
+            tempBuffer[outIdx++] = ((double)inClose[today] - lowest) / diff;
          } else {
             tempBuffer[outIdx++] = 0.0;
          }
